backend/requirements.txt
# ProtoQol Enterprise Backbone Dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
aiohttp==3.9.1
httpx==0.26.0
google-generativeai==0.3.2
solders==0.18.0
solana==0.30.2
anchorpy==0.19.0
python-dotenv==1.0.0
aiogram==3.3.0
gspread==6.0.0
google-auth==2.27.0
